requirements.txt
spaces
torch == 2.4.0
torchvision
torchaudio
python-dotenv
cython
gitpython >= 3.1
tensorboard >= 2.11
numpy >= 1.21, <2.1
Pillow >= 9.5
opencv-python >= 4.8
scipy >= 1.7
tqdm >= 4.66.1
gradio >= 3.34
einops >= 0.6
hydra-core >= 1.3.2
requests
torchdiffeq
librosa >= 0.8.1
nitrous-ema
safetensors
auraloss
hydra_colorlog
tensordict
colorlog
open_clip_torch
soundfile
av
transformers
ffmpeg-python
moviepy
scenedetect
scikit-learn == 1.6.1
qwen-vl-utils[decord]
openai
accelerate